feat: Many many changes

+++ src/module/core/texops.py
@@ -37,7 +37,7 @@ def normalize(img: Image, size: tuple[int, int]|None=None, mode: Literal['L', 'R
 	if size:
 		img = img.resize(size)
 
-	img = img.convert(np.float16)
+	img = img.convert(np.float32)
 
 	if mode:
 		img = img.normalize(mode)
@@ -45,7 +45,6 @@ def normalize(img: Image, size: tuple[int, int]|None=None, mode: Literal['L', 'R
 		img = img.normalize('RGB')
 
 	return img
-
 
 def make_phong_exponent(mat: Material) -> Image:
 	''' Generates an RGB phong exponent texture. '''
@@ -141,6 +140,7 @@ def make_basecolor(mat: Material) -> Image:
 	# Envmap mask as basetexture alpha
 	elif not using_phong and MaterialMode.embed_envmap(mat.mode):
 		# TODO: This sucks, but srctools has forced my hand. Libsquish needs a flag to account for full-alpha, which we can't give it.
+		# TODO: Verify that this is still an issue with sourcepp bcenc?
 		envmask = make_envmask(mat)
 		envmask.data = envmask.data.clip(min=(1 / 255))
 		return Image.merge((r, g, b, envmask))
